Edit docstrings for distributions
Also automatically add attributes with @metadist
