+++ metasyn/provider.py
@@ -20,15 +20,8 @@ import polars as pl
 
 from metasyn.distribution.base import BaseDistribution
 from metasyn.distribution.categorical import MultinoulliDistribution
-from metasyn.distribution.constant import (
-    ConstantDistribution,
-    DateConstantDistribution,
-    DateTimeConstantDistribution,
-    DiscreteConstantDistribution,
-    StringConstantDistribution,
-    TimeConstantDistribution,
-)
 from metasyn.distribution.continuous import (
+    ConstantDistribution,
     ExponentialDistribution,
     LogNormalDistribution,
     NormalDistribution,
@@ -36,24 +29,30 @@ from metasyn.distribution.continuous import (
     UniformDistribution,
 )
 from metasyn.distribution.datetime import (
+    DateConstantDistribution,
+    DateTimeConstantDistribution,
     DateTimeUniformDistribution,
     DateUniformDistribution,
+    TimeConstantDistribution,
     TimeUniformDistribution,
 )
 from metasyn.distribution.discrete import (
+    DiscreteConstantDistribution,
     DiscreteNormalDistribution,
     DiscreteTruncatedNormalDistribution,
     DiscreteUniformDistribution,
     PoissonDistribution,
     UniqueKeyDistribution,
 )
-from metasyn.distribution.faker import (
+from metasyn.distribution.na import NADistribution
+from metasyn.distribution.string import (
     FakerDistribution,
     FreeTextDistribution,
+    RegexDistribution,
+    StringConstantDistribution,
     UniqueFakerDistribution,
+    UniqueRegexDistribution,
 )
-from metasyn.distribution.na import NADistribution
-from metasyn.distribution.regex import RegexDistribution, UniqueRegexDistribution
 from metasyn.privacy import BasePrivacy, BasicPrivacy
 from metasyn.util import get_registry
 from metasyn.varspec import DistributionSpec
